% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2parsList.R
\name{add2parsList}
\alias{add2parsList}
\title{add2parsList}
\usage{
add2parsList(oneList, object)
}
\arguments{
\item{oneList}{List in which the new parameter will be introduced.}

\item{object}{Object to add in the list.}
}
\value{
Returns the list with the new parameter.
}
\description{
add2parsList
}
