% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToLinksFile.R
\name{addToLinksFile}
\alias{addToLinksFile}
\title{addToLinksFile}
\usage{
addToLinksFile(linksFile, aFName, categ = "", desc = "", subcateg = "")
}
\arguments{
\item{linksFile}{Name of the links file.}

\item{aFName}{Name of the new object to register in the links file.}

\item{categ}{Category of the new file.}

\item{desc}{Description of the file.}

\item{subcateg}{Subcategory of the file.}
}
\description{
addToLinksFile
}
