% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults2Browser.R
\name{LinksFile2Html}
\alias{LinksFile2Html}
\title{A Function to create HTML from file of links}
\usage{
LinksFile2Html(lFile, outputDir, info.list, categs.descs = NULL,
  categs.names = NULL, IndexDir = "", UEB = TRUE,
  resultsFileName = "ResultFiles")
}
\arguments{
\item{lFile}{Input file consisting of 4 tab separated fields: Filename, Categ, Subcateg, Description}

\item{outputDir}{Name of the directory where the HTML is to be written}

\item{info.list}{List with informations to build the header: To, Description, Analysts, Contact}

\item{categs.descs}{Description of categories in the links file (Header of sections)}

\item{categs.names}{Short names for categories in the links file (to be used in the input file)}

\item{IndexDir}{Name of subdirectory where the files to be linked are located}

\item{UEB}{TRUE if UEB header is to be used}

\item{resultsFileName}{Name of output file}
}
\description{
This function allows you to create an HTML from a text file with a predefined structure
}
\examples{
linksFileName <-"LinksFileName_Example.csv"
resultsDir <- "."
# UEB <- TRUE # if UEB == TRUE => UEB header is used
htmlInfo <- list(To = "My Client", # nomClients,
                 Description = "Analysis of Biological Significance for some selected gene list",
                 Analysts = "Ferran Brianso and Alex Sanchez",               # analistes,
                 Contact = "Alex Sanchez (alex.sanchez@vhir.org)")
GENCategs <- myGENCategs() 
fcategs <- GENCategs[["categs"]]
fNames <-  GENCategs[["noms"]]
# Invoke main function 
# Not run automatically. Check by yourself uncommenting next lines
# setwd("examples")
# LinksFile2Html(linksFileName, resultsDir, htmlInfo, 
#                categs.descs = fCategs, categs.names = fNames,
#                IndexDir = "", UEB = TRUE, resultsFileName="Resultats")
}
\keyword{HTML,}
\keyword{links}

