% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults2Browser.R
\name{myFACategs}
\alias{myFACategs}
\title{myFACategs}
\usage{
myFACategs()
}
\description{
This function returns names and descriptions for a 'Functional Analysis' results file
}
\examples{
myCategs <-myFACategs()
categorias <- myCategs[["categs"]]
nombres <-  myCategs[["noms"]]
}
\keyword{categories}

