% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults2Browser.R
\name{myGENCategs}
\alias{myGENCategs}
\title{myGENCategs}
\usage{
myGENCategs()
}
\description{
This function returns names and descriptions for a general results file
}
\examples{
myCategs <-myGENCategs()
categorias <- myCategs[["categs"]]
nombres <-  myCategs[["noms"]]
}
\keyword{categories}

